GH-3907: reduce the Marten and Polecat aggregate attributes to shells
Both integrations now inherit the core vocabulary instead of reimplementing it:

  Wolverine.Marten.WriteAggregateAttribute   : WriteModelAttribute
  Wolverine.Marten.ReadAggregateAttribute    : ReadModelAttribute
  Wolverine.Marten.AggregateHandlerAttribute : DeciderFunctionAttribute

and the same three in Wolverine.Polecat. This is purely additive - every
existing spelling still compiles and generates the same code. What went away is
the second and third copy of the workflow behind them: ~1,000 lines across
AggregateHandling, MissingAggregateCheckFrame and TagAggregateOtelFrame in the
two integrations, which increment 1's drift reconciliation had already brought
to differing only by the store's name.

Two of the six are literally empty - [ReadAggregate] on both stores is nothing
but its two constructors. The other four are not, and for one reason worth
recording: ConcurrencyStyle is a public enum in each integration's own
namespace, so [WriteAggregate(LoadStyle = ConcurrencyStyle.Exclusive)] and
[AggregateHandler(ConcurrencyStyle.Exclusive)] resolve it to the store's
spelling. Retiring those enums would break that source, and GH-3907 retires
nothing, so each store keeps its enum as public vocabulary - the same call
already made for Events and UnknownAggregateException - and the shell bridges
it: one shadowing property on [WriteAggregate], one forwarding constructor on
[AggregateHandler]. The two enums are the same two members in the same order.

The seam is implemented ON MartenPersistenceFrameProvider and
PolecatPersistenceFrameProvider rather than on the standalone provider classes
increment 1 created. That is what makes it discoverable: the workflow resolves
the owning store out of the persistence strategies already registered on
GenerationRules - the same registry [Entity] resolves through - so nothing in
Wolverine core enumerates stores. The standalone classes are deleted.

Each store's DetermineSagaIdType turned out to be exactly the identity
resolution its WriteAggregateAttribute had inlined - Marten resolving the
document type's IdType, Polecat reflecting over the Id property - so that seam
already existed and the shared workflow just calls it.

FetchLatestAggregateFrame moved out of ReadAggregateAttribute.cs into each
store's Codegen folder, unchanged. Nothing names it now; core reaches it
through BuildFetchLatestFrame, which is what keeps Marten's IBatchableFrame
enlistment - a Marten-only extra Polecat has no equivalent of - private to
Marten.

Wolverine.Http.Marten and .Polecat read AggregateHandling off the chain to
describe the workflow to OpenAPI. The tag key is unchanged, so they only needed
the new namespace.
