+++ src/Persistence/Wolverine.Polecat/BoundaryModelAttribute.cs
@@ -16,6 +16,7 @@ using Wolverine.Polecat.Persistence.Sagas;
 using Wolverine.Persistence;
 using Wolverine.Runtime;
 using Wolverine.Runtime.Handlers;
+using Wolverine.Persistence.EventSourcing;
 
 namespace Wolverine.Polecat;
